Nabegazio menua Fragment-ekin eginda. Guneak fragment, Profila fragment eta Itxi saioa fragmentetara pasatzen da.
