+++ README.md
@@ -1,4 +1,4 @@
 # LaudioSarean
 
-Talde 6 (Euskera)  
-Mikel Monasterio Alday, Aritza Herrero Niño, Unax Zulaika Fuente, Aingeru Siranaula Santos
+Talde 6 (Euskera)
+Aritza Herrero Niño, Unax Zulaika Fuente, Aingeru Siranaula Santos
