+++ app/src/main/res/layout/fragment_itxi_saioa.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".ItxiSaioaFragment">
+
+    <!-- TODO: Update blank fragment layout -->
+    <TextView
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:text="Itxi Saioa"
+        android:layout_marginTop="30dp"
+        android:gravity="center_horizontal"
+        android:textSize="30sp"/>
+
+</FrameLayout>
