+++ app/src/main/res/values/strings.xml
@@ -41,4 +41,6 @@
     <string name="izenburuaGunea"></string>
     <string name="informazioGunea"></string>
     <string name="play">Jolasa hasi</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
