+++ app/src/main/res/values/themes.xml
@@ -3,9 +3,18 @@
     <style name="Base.Theme.LaudioSarean" parent="Theme.Material3.DayNight.NoActionBar">
         <!-- Customize your light theme here. -->
         <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
+        <item name="colorPrimary">@color/green_normal</item>
+        <item name="colorSecondary">@color/green_dark</item>
         <item name="android:statusBarColor">@color/white</item>
         <item name="android:windowLightStatusBar">true</item>
     </style>
 
     <style name="Theme.LaudioSarean" parent="Base.Theme.LaudioSarean" />
+
+    <style name="BottomNavigation">
+        <item name="android:background">@color/green_dark</item>
+        <item name="itemIconTint">@color/green_light2</item>
+        <item name="itemTextColor">@color/green_light2</item>
+
+    </style>
 </resources>
